Create trust helpers library
